Separate io_u from ioengine [1/3] - add io_u.h
Extract io_u related structures and functions from ioengine.h to
a new header io_u.h which is intended to be a header for io_u.c
(similar to the way there is io_u_queue.h for io_u_queue.c),
since there isn't much reason for them to be a part of ioengine.h.

This commit does nothing other than separating them out (and make
other related files include the new header).

Confirmed this compiles on Linux, Cygwin and several BSDs.
